Fix ignored AOV clear FramePass parameter.
The clear flag was ignored, unless AOV was visualized. The clear logic
was also simplified by removing the call to ResetRenderOutputClear()
from RenderBufferManager::Impl::SetRenderOutputs(). This was the old
way to clear the render buffers and is no more necessary, since
RenderBufferManager::SetRenderOutputClearColor() can now be called with
an empty VtValue instead, which will yield the same result and reset
the clear value. Furthermore, it is preferable to set the clear value
outside of RenderBufferManager::Impl::SetRenderOutputs(), which already
contains too much complexity.