OGSMOD-8007: Disable Skydome on Linux with MSAA off.
Unit tests TestMsaaAAOff and TestMsaaWireframeAAOff currently have a failure rate of around 15-20% with Linux release builds.
At first, we thought it was only happening in debug because we did not catch it right away, but it also happens in release.
The associated Jira issue OGSMOD-8007 was opdated with than new information.