OGSMOD-8106: Fix unstable TestFramePasses unit tests (#82)
Wait for GPU work to be completed on the 1st frame pass before executing
the 2nd one for the following unit test:
- TestFramePasses_MultiViewports
- TestFramePasses_ClearColorBuffer
- TestFramePasses_TestDynamicAovInputs

This improves the success rate with Metal, where the render pass state
data (camera matrices, lighting) is sometimes corrupted/incorrect.

This is a temporary solution, waiting for the GPU to be idle before starting
the 2nd frame pass is inefficient, so we will have to implement a better sync
strategy. Until then, this provides more unit test stability.